requirements.txt
numpy~=1.26.4
torch~=2.2.0
python~=3.12.1
matplotlib==3.8.2
gpytorch==1.11
scikit-learn==1.4.0
pennylane==0.34.0
qiskit<0.46
omegaconf==2.3.0
hydra-core==1.3.2
ax-platform==0.3.6
scipy==1.12.0
pandas==2.2.1